+++ examples/demo-viewer/agent/demo/langgraph_agentic_chat/page.tsx
@@ -3,12 +3,13 @@ import React, { useState } from "react";
 import "@copilotkit/react-ui/styles.css";
 import "./style.css";
 import { CopilotKit, useCopilotAction } from "@copilotkit/react-core";
-import { CopilotChat } from "@copilotkit/react-ui";
-
+import { CopilotChat, useCopilotChatSuggestions } from "@copilotkit/react-ui";
+import { chatSuggestions, initialPrompt } from "@/lib/prompts";
 const AgenticChat: React.FC = () => {
   return (
     <CopilotKit
-      publicApiKey={process.env.NEXT_PUBLIC_COPILOT_CLOUD_API_KEY}
+      // publicApiKey={process.env.NEXT_PUBLIC_COPILOT_CLOUD_API_KEY}
+      runtimeUrl="/api/copilotkit"
       showDevConsole={false}
       agent="agentic_chat"
     >
@@ -36,6 +37,10 @@ const Chat = () => {
     },
   });
 
+  useCopilotChatSuggestions({
+    instructions: chatSuggestions.agenticChat,
+  })
+
   return (
     <div
       className="flex justify-center items-center h-full w-full"
@@ -44,7 +49,7 @@ const Chat = () => {
       <div className="w-8/10 h-8/10 rounded-lg">
         <CopilotChat
           className="h-full rounded-lg"
-          labels={{ initial: "Hi, I'm an agent. Want to chat?" }}
+          labels={{ initial: initialPrompt.agenticChat }}
         />
       </div>
     </div>