+++ examples/demo-viewer/agent/demo/langgraph_agentic_generative_ui/page.tsx
@@ -3,12 +3,13 @@ import React from "react";
 import "@copilotkit/react-ui/styles.css";
 import "./style.css";
 import { CopilotKit, useCoAgentStateRender } from "@copilotkit/react-core";
-import { CopilotChat } from "@copilotkit/react-ui";
-
+import { CopilotChat, useCopilotChatSuggestions } from "@copilotkit/react-ui";
+import { initialPrompt, chatSuggestions } from "@/lib/prompts";  
 const AgenticGenerativeUI: React.FC = () => {
   return (
     <CopilotKit
-      publicApiKey={process.env.NEXT_PUBLIC_COPILOT_CLOUD_API_KEY}
+      // publicApiKey={process.env.NEXT_PUBLIC_COPILOT_CLOUD_API_KEY}
+      runtimeUrl="/api/copilotkit"
       showDevConsole={false}
       agent="agentic_generative_ui"
     >
@@ -69,13 +70,15 @@ const Chat = () => {
       );
     },
   });
-
+  useCopilotChatSuggestions({
+    instructions: chatSuggestions.agenticGenerativeUI,
+  })
   return (
     <div className="flex justify-center items-center h-screen w-screen">
       <div className="w-8/10 h-8/10">
         <CopilotChat
           className="h-full rounded-lg"
-          labels={{ initial: "Hi, I'm an agent. I can do anything, just ask!" }}
+          labels={{ initial: initialPrompt.agenticGenerativeUI }}
         />
       </div>
     </div>