+++ examples/demo-viewer/agent/demo/langgraph_human_in_the_loop/page.tsx
@@ -3,13 +3,13 @@ import React, { useState, useEffect } from "react";
 import "@copilotkit/react-ui/styles.css";
 import "./style.css";
 import { CopilotKit, useLangGraphInterrupt } from "@copilotkit/react-core";
-import { CopilotChat } from "@copilotkit/react-ui";
-
+import { CopilotChat, useCopilotChatSuggestions } from "@copilotkit/react-ui";
+import { initialPrompt, chatSuggestions } from "@/lib/prompts";  
 const HumanInTheLoop: React.FC = () => {
   return (
     <CopilotKit
-      publicApiKey={process.env.NEXT_PUBLIC_COPILOT_CLOUD_API_KEY}
-      // runtimeUrl="/api/copilotkit"
+      // publicApiKey={process.env.NEXT_PUBLIC_COPILOT_CLOUD_API_KEY}
+      runtimeUrl="/api/copilotkit"
       showDevConsole={false}
       agent="human_in_the_loop"
     >
@@ -88,42 +88,19 @@ const Chat = () => {
       );
     },
   });
-
+  useCopilotChatSuggestions({
+    instructions: chatSuggestions.humanInTheLoop,
+  })
   return (
     <div className="flex justify-center items-center h-screen w-screen">
       <div className="w-8/10 h-8/10">
         <CopilotChat
           className="h-full rounded-lg"
-          labels={{ initial: "Hi, I'm an agent. I can do anything, just ask!" }}
+          labels={{ initial: initialPrompt.humanInTheLoop }}
         />
       </div>
     </div>
   );
 };
-
-function Spinner() {
-  return (
-    <svg
-      className="mr-2 size-3 animate-spin text-slate-500"
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-    >
-      <circle
-        className="opacity-25"
-        cx="12"
-        cy="12"
-        r="10"
-        stroke="currentColor"
-        strokeWidth="4"
-      ></circle>
-      <path
-        className="opacity-75"
-        fill="currentColor"
-        d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-      ></path>
-    </svg>
-  );
-}
 
 export default HumanInTheLoop;