+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @copilotkit/runtime-client-gql
 
+## 1.8.10
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.10
+
+## 1.8.10-next.3
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.10-next.3
+
+## 1.8.10-next.2
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.10-next.2
+
+## 1.8.10-next.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.10-next.1
+
+## 1.8.10-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.10-next.0
+
 ## 1.8.9
 
 ### Patch Changes